requirements.txt
fastapi
asyncpg
uvicorn
psycopg2-binary
strawberry-graphql